Made updates to the skeleton of backend.
