+++ models/algorithms_utility.py
@@ -1,0 +1,4 @@
+class Algorthms:
+    
+    def __init__():
+        pass
