--- models/main_model.py
@@ -1,117 +1,0 @@
-# from typing_extensions import Required
-from flask import request
-from flask_restful import Resource, reqparse
-
-import osmnx
-# import requests
-# import time
-
-import networkx as nx
-# import pandas as pd
-
-
-class MainModel(Resource):
-
-    parser = reqparse.RequestParser()
-    parser.add_argument('source',
-                        type=str,
-                        required=True,
-                        help="This field cannot be blank."
-                        )
-    parser.add_argument('destination',
-                        type=str,
-                        required=True,
-                        help="This field cannot be blank."
-                        )
-    parser.add_argument('is_elevation_max',
-                        type=str,
-                        required=True,
-                        help="This field cannot be blank."
-                        )
-    parser.add_argument('percentage',
-                        type=str,
-                        required=True,
-                        help="This field cannot be blank."
-                        )
-    parser.add_argument('algorithm',
-                        type=str,
-                        required=True,
-                        help="This field cannot be blank."
-                        )
-
-
-    def post(self):
-        data = self.parser.parse_args()
-        
-
-
-        source = "umass police station"
-        destintion = "puffton village apartments"
-
-        source_coord = osmnx.geocoder.geocode(source)
-        destination_coord = osmnx.geocoder.geocode(destintion)
-        is_max = True
-        percentage = 50
-        print(source_coord)
-        print(destination_coord)
-
-        params = dict()
-        params['city']  = "Amherst"
-        params['state'] = "Massachusetts"
-        params['country'] = 'USA'
-        graph = osmnx.graph_from_place(params, network_type='drive')
-        # graph = self.add_node_elevations_open(G=graph)
-        # call methods to 
-        # gen map
-        # find node ids
-        # add elevation
-        # call algo
-
-        return {'elevation_route': 
-                {'properties': {}, 
-                'type': 'Feature', 
-                'geometry': 
-                {'type': 'LineString', 
-                'coordinates': [[-122.4510285, 37.7842783], 
-                [-122.4501543, 37.784488], 
-                [-122.4492965, 37.7846918], 
-                [-122.4483551, 37.7849203], 
-                [-122.4480059, 37.7850101], 
-                [-122.447628, 37.7850954], 
-                [-122.4472443, 37.7851849], 
-                [-122.4468749, 37.7852722], 
-                [-122.4463795, 37.7850871], 
-                [-122.4462556, 37.7844446], 
-                [-122.4460465, 37.7835187], 
-                [-122.4458607, 37.7826171], 
-                [-122.4458163, 37.7824025], 
-                [-122.4457431, 37.7824089], 
-                [-122.4450799, 37.7823361], 
-                [-122.4448947, 37.7823589], 
-                [-122.4443955, 37.7819342], 
-                [-122.4441933, 37.7819698], [-122.4441687, 37.7818063], [-122.4438512, 37.7818735], 
-                [-122.4437725, 37.7818859], [-122.4427013, 37.7819811], [-122.4422579, 37.7820536], 
-                [-122.4422813, 37.7821202], [-122.442126, 37.7821569], [-122.4418307, 37.7822406], 
-                [-122.4416383, 37.7822905], [-122.4415744, 37.7821603], [-122.4409913, 37.7822555], 
-                [-122.4400565, 37.7823643], [-122.4399177, 37.7823828], [-122.4393709, 37.7824554], 
-                [-122.439275, 37.7824675], [-122.4384796, 37.7825679], [-122.437655, 37.7826719], 
-                [-122.4375511, 37.7821576], [-122.4374661, 37.7817369], [-122.4373847, 37.7813336], 
-                [-122.4372786, 37.7808087], [-122.4371853, 37.7803465], [-122.4370897, 37.7798739], 
-                [-122.4362387, 37.7799822], [-122.4354442, 37.7800834], [-122.4352234, 37.7801115], 
-                [-122.4343145, 37.7802272], [-122.434279, 37.7802317], [-122.4338018, 37.7802924], [-122.4337144, 37.7798606], [-122.4335103, 37.7798862], [-122.4332403, 37.7799201], [-122.4331295, 37.7794214], [-122.4327217, 37.7794742], [-122.4326755, 37.7794801], [-122.4319694, 37.779569], [-122.4314941, 37.7796294], [-122.4309983, 37.7796924], [-122.4309337, 37.7797006], [-122.4303243, 37.7797781], [-122.4301795, 37.7797965], [-122.429533, 37.7798832], [-122.4291361, 37.7799313], [-122.4287655, 37.7799764], [-122.4285898, 37.7799987], [-122.4285392, 37.7797423], [-122.428504, 37.7795627], [-122.4284019, 37.7790656], [-122.4283549, 37.7788409], [-122.4285348, 37.7788149], [-122.4284216, 37.778596], [-122.4283826, 37.778597], [-122.4282844, 37.7783612], [-122.4282158, 37.7781354], [-122.4277533, 37.7781941], [-122.4274483, 37.7782327], [-122.4272268, 37.7782608], [-122.426662, 37.7783324], [-122.4265671, 37.7778687], [-122.4264716, 37.7774027], [-122.4263787, 37.7769363], [-122.4262857, 37.7764706], [-122.4261917, 37.7760035], [-122.4260983, 37.7755395], [-122.4260018, 37.7750597], [-122.4259102, 37.7746045], [-122.4258171, 37.7741418], [-122.4257227, 37.7736729], [-122.4256255, 37.7732088], [-122.4255299, 37.7727425], [-122.4254662, 37.7724228], [-122.4253435, 37.7718028], [-122.4252937, 37.7715638], [-122.4251529, 37.7708881], [-122.4249951, 37.7707352], [-122.4248322, 37.7706553], [-122.4246666, 37.7700159], [-122.4246367, 37.7697392], [-122.4248662, 37.7697255], [-122.4250784, 37.7697129], [-122.4255084, 37.7696873], [-122.4255844, 37.7696828], [-122.4257494, 37.7696729], [-122.4261321, 37.7696502], [-122.4264634, 37.7693878], [-122.4266615, 37.7692309], [-122.4267308, 37.7691759], [-122.4267436, 37.7689979], [-122.4266919, 37.7684598], [-122.4266363, 37.7678819], [-122.4265571, 37.7670573], [-122.4264815, 37.7662713], [-122.4264346, 37.7657848], [-122.42639, 37.765319], [-122.4263255, 37.7646472], [-122.4262902, 37.7643325], [-122.4262793, 37.7642253], [-122.4262613, 37.7639799], [-122.42622, 37.7635498], [-122.4262172, 37.7635221], [-122.42617, 37.7630289], [-122.4261318, 37.7626322], [-122.4260895, 37.7622008], [-122.4262355, 37.7621616], [-122.4262301, 37.7620957], [-122.4261652, 37.7614218], [-122.4261593, 37.7613602], [-122.4262605, 37.7613541], [-122.4262196, 37.7609278], [-122.4261135, 37.7598191], [-122.4260111, 37.7598192], [-122.4258623, 37.7598278], [-122.4248801, 37.7598932], [-122.42369, 37.759966], [-122.4235989, 37.7599724], [-122.422879, 37.760013]]}}, 
-                'shortest_route': {
-                    'properties': {}, 
-                    'type': 'Feature', 
-                    'geometry': 
-                    {'type': 'LineString', 
-                    'coordinates': [[-122.4510285, 37.7842783], [-122.4501543, 37.784488], [-122.4501223, 37.7821812], [-122.4500984, 37.7820602], [-122.4498377, 37.7807564], [-122.4495522, 37.7807927], [-122.4493128, 37.7808231], [-122.449032, 37.7804186], [-122.4482027, 37.7800573], [-122.4473713, 37.7801607], [-122.4472255, 37.7794289], [-122.4471331, 37.778964], [-122.4470623, 37.7786307], [-122.4468691, 37.777674], [-122.446681, 37.776743], [-122.4464925, 37.7758097], [-122.4463114, 37.7748781], [-122.4462413, 37.774476], [-122.4461264, 37.7739463], [-122.4459907, 37.7733251], [-122.445924, 37.7729945], [-122.4457704, 37.7722293], [-122.445742, 37.7720935], [-122.4455504, 37.7711507], [-122.4453607, 37.7702176], [-122.4451431, 37.7692913], [-122.4447508, 37.7674251], [-122.4445557, 37.7664337], 
-                    [-122.444172, 37.765959], [-122.4434553, 37.7653549], [-122.4427054, 37.7653393], [-122.4425773, 37.7652063], [-122.442386, 37.7652536], [-122.4423133, 37.7650093], [-122.439457, 37.7636301], [-122.4390295, 37.7634605], [-122.4360944, 37.7634194], [-122.4352756, 37.76347], [-122.4352475, 37.7631776], [-122.4352156, 37.7628445], [-122.4350308, 37.7628033], [-122.4349699, 37.7627231], [-122.4349753, 37.7624994], [-122.434064, 37.7625548], [-122.4329647, 37.7626275], [-122.4321262, 37.7626773], [-122.4315473, 37.7627135], [-122.4307266, 37.7627623], [-122.4306733, 37.7620781], [-122.4299843, 37.7620816], [-122.4284491, 37.762071], [-122.4284481, 37.7620605], [-122.4283911, 37.7612812], [-122.4282288, 37.7612974], [-122.4282018, 37.7612365], [-122.4281383, 37.7612405], [-122.4280932, 37.7611458], [-122.4279349, 37.7607795], [-122.4278854, 37.7607822], [-122.4277969, 37.7605617], [-122.4267964, 37.7598573], [-122.4264606, 37.7598442], [-122.4264564, 37.7597992], [-122.4263723, 37.7598041], [-122.4262145, 37.7598132], [-122.4261135, 37.7598191], [-122.4260111, 37.7598192], [-122.4258623, 37.7598278], [-122.4248801, 37.7598932], [-122.42369, 37.759966], [-122.4235989, 37.7599724], [-122.422879, 37.760013]]}}, 
-                    'shortDist': 4780.319000000001, 
-                    'gainShort': 104.61599999999997, 
-                    'dropShort': 176.31499999999997, 
-                    'elenavDist': 5598.104999999999, 
-                    'gainElenav': 66.34300000000003, 
-                    'dropElenav': 138.04200000000003, 
-                    'popup_flag': 2}
-
-
