+++ models/map_utility.py
@@ -1,0 +1,5 @@
+class Map:
+    
+    def __init__():
+        pass
+
